Move trace result to OPA output channel
There is no need to have a dedicated provider for the trace as it exists
today. In the future we can improve the tracing output, which may
require a custom provider, but as is, just send it to the output
channel.

Also, use the 'pretty' format instead of 'json' which now prints the
trace events in a machine-readable format.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>