Update partial eval command to show picklist
Previously the extension did not let the user select unknowns. With
these changes, the extension will display a list of base document
dependencies that the user can select from.

In the future, we may want to include more options (e.g., the transitive
closure of refs), support for multiple options, or support for arbitrary
unknowns.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>