Update auto check to use entire workspace
Previously the auto-check was only passing the current file to the check
subcommand. As a result, if multiple files contributed to the same
package and depended on each other, the check subcommand would raise an
error. These changes just fix the plugin to pass the root path to the
check subcommand so that the entire workspace is rechecked on each save.

If this becomes too expensive we can investigate improvements like only
rechecking files that have changed and their dependencies (though this
would require new static analysis.)

Fixes #6

Signed-off-by: Torin Sandall <torinsandall@gmail.com>