Fix getInputPath to use Uri.file instead of Uri.parse
On Windows, the Uri.parse function was returning a URI object that was
not understood by the rest of the extension--the URI object would have
the scheme set to 'c' instead of 'file'. As a result, none of the URI
handling was working and the extension was falling back to the
path.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>