fmt/goto-defn: drop functionality in favor of LSP
We are trying to standardize Rego developer env functionality
on the Regal LS implementation to be able to offer a standardized
experience to all Rego developers, regardless of their preferred editor.

Fmt on save and goto definition are now available as part of the
language server and so should be accessed via that route instead.

Signed-off-by: Charlie Egan <charlie@styra.com>