Adapt to new `print` output map structure
Regal now correctly sends back print output not just by **line** but as
**file -> line**. The previous assumption that only lines from the
current file would be returned was wrong of course, and could lead to
print output from other files getting displayed in the editor.

Ideally we'd be able to decorate lines in any file where print output
is provided, but for now, we'll just show it in the active editor window.
In order to still be able to see print output from other files, we'll send
the output of all to the Regal output channel.

Signed-off-by: Anders Eknert <anders@eknert.com>